Fix handing of sequence number on TCP transmit
The variable _sckt->timeout is used for the sole purpose of
determining whether the sequence number should be incremented, or
whether this is a retransmit.  However, the updating of this variable
didn't really work correctly since nwk_upstream can be called through
other means than from nwk_tick (there is for example a "Reschedule
50ms later for eventual further processing") which is where the
variable would be updated.  The result being that the sequence number
would be incorrectly incremented on any retransmit caused by a call to
nwk_upstream not coming from nwk_tick.

Change this to the more robust approach of setting timeout whenever
the sequence number is incremented, and clearing it only when there
is new data to transmit.

This variable could probably use a new name as well.  :-)